Dedupe breadcrumb
